Added States to the Monster AI
Added a chasing, patrolling and disabled state and logic to switch between them. The chasing state is the old behavior. The patrolling state has the monster follow an ever growing path around the player for as long as the player's position is unreachable. The disabled state prevents the monster from preforming any actions (this will be used to stop the monster when the lights go out.)