Prefab for Switches
The switches now have a prefab with all of the scripts, positions, colliders, and lamp built into it.  The spawn script has been modified to use this prefab.

NOTE: In order for the lights to be correctly oriented the blue local vector needs to be facing right along the wall, and from top orthographic view the point needs to be lined up with the outermost line of the wall.